[otbn] Move temporary directory creation into the ISS wrapper
This is part of moving to "make an ISS wrapper; prod it; get results".
The first and last steps need the temporary directory, so it seems
sensible to bundle it up with the ISS wrapper.
