+++ hw/ip/otbn/dv/model/otbn_model.cc
@@ -3,16 +3,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <cassert>
-#include <cstdlib>
-#include <cstring>
 #include <fstream>
-#include <ftw.h>
 #include <iostream>
-#include <memory>
 #include <sstream>
 #include <string>
 #include <svdpi.h>
-#include <sys/stat.h>
 
 #include "iss_wrapper.h"
 
@@ -120,88 +115,6 @@ static void load_memory(const std::string &path, const char *scope,
   }
 }
 
-// Guard class to create (and possibly delete) temporary directories.
-struct TmpDir {
-  std::string path;
-
-  ~TmpDir() { cleanup(); }
-
-  // Construct a temporary directory. Throw a runtime_error if something goes
-  // wrong.
-  void make() {
-    // Clean up anything we had already.
-    cleanup();
-    path = TmpDir::make_tmp_dir();
-  }
-
- private:
-  // A wrapper around mkdtemp that respects TMPDIR
-  static std::string make_tmp_dir() {
-    const char *tmpdir = getenv("TMPDIR");
-    if (!tmpdir)
-      tmpdir = "/tmp";
-
-    std::string tmp_template(tmpdir);
-    tmp_template += "/otbn_XXXXXX";
-
-    if (!mkdtemp(&tmp_template.at(0))) {
-      std::ostringstream oss;
-      oss << ("Cannot create temporary directory for OTBN simulation "
-              "with template ")
-          << tmp_template << ": " << strerror(errno);
-      throw std::runtime_error(oss.str());
-    }
-
-    // The backing string for tmp_template will have been populated by mkdtemp.
-    return tmp_template;
-  }
-
-  // Return true if the OTBN_MODEL_KEEP_TMP environment variable is set to 1.
-  static bool should_keep_tmp() {
-    const char *keep_str = getenv("OTBN_MODEL_KEEP_TMP");
-    if (!keep_str)
-      return false;
-    return (strcmp(keep_str, "1") == 0) ? true : false;
-  }
-
-  // Called by nftw when we're deleting the temporary directory
-  static int ftw_callback(const char *fpath, const struct stat *sb,
-                          int typeflag, struct FTW *ftwbuf) {
-    // The libc remove() function calls unlink or rmdir as necessary. Ignore
-    // any failures: we'll check that we managed to delete the directory when
-    // nftw finishes.
-    remove(fpath);
-
-    // Tell nftw to keep going
-    return 0;
-  }
-
-  // Recursively delete the temporary directory
-  void cleanup() {
-    if (path.empty())
-      return;
-
-    if (TmpDir::should_keep_tmp()) {
-      std::cerr << "Keeping temporary directory at " << path
-                << " because OTBN_MODEL_KEEP_TMP=1.\n";
-      return;
-    }
-
-    // We're not supposed to keep the directory. Try to delete it and its
-    // contents. Ignore any failures: we'll just check whether it's gone
-    // afterwards.
-    nftw(path.c_str(), TmpDir::ftw_callback, 4, FTW_DEPTH | FTW_PHYS);
-
-    // Is there still anything at path? If so, we failed. Print something to
-    // stderr to tell the user what's going on.
-    struct stat statbuf;
-    if (stat(path.c_str(), &statbuf) == 0) {
-      std::cerr << "ERROR: Failed to delete OTBN temporary directory at "
-                << path << ".\n";
-    }
-  }
-};
-
 extern "C" ISSWrapper *otbn_model_init() {
   try {
     return new ISSWrapper();
@@ -221,17 +134,8 @@ extern "C" int otbn_model_run(ISSWrapper *model, const char *imem_scope,
   assert(dmem_words >= 0);
   assert(start_addr >= 0 && start_addr < (imem_words * 4));
 
-  TmpDir tmpdir;
-  try {
-    tmpdir.make();
-  } catch (const std::runtime_error &err) {
-    std::cerr << err.what() << "\n";
-    return 0;
-  }
-
-  std::string ifname(tmpdir.path + "/imem");
-  std::string dfname(tmpdir.path + "/dmem");
-  std::string tfname(tmpdir.path + "/trace");
+  std::string ifname(model->make_tmp_path("imem"));
+  std::string dfname(model->make_tmp_path("dmem"));
 
   try {
     dump_memory(dfname, dmem_scope, dmem_words, 32);
